requirements/base.txt
alembic==1.7.7
aniso8601==9.0.1
Babel==2.10.1
blinker==1.4
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.1
Flask==2.1.2
Flask-Alembic==2.0.1
Flask-BabelEx==0.9.4
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.36
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
gunicorn==20.1.0
google-auth==2.6.6